refactor: import ACTIVITY_TYPES directly from shared/ in ApprovalsTimeline
