+++ backend/src/services/approvalService.js
@@ -28,7 +28,7 @@
 
 import * as testRepo from "../database/repositories/testRepo.js";
 import * as activityRepo from "../database/repositories/activityRepo.js";
-import { ACTIVITY_TYPES } from "../utils/activityTypes.js";
+import { ACTIVITY_TYPES } from "../../../shared/activityTypes.js";
 
 /** `tests.approvalSource` enumeration. */
 export const APPROVAL_SOURCE = Object.freeze({
